refactor(suite-native): fold earn presets into components/HowEarnWorks
