+++ suite-native/module-earn/src/components/HowEarnWorks/stakePresets.tsx
